Add Create Product FAQ Controller and display FAQ List on Page

+++ HelloQQPortal/Database/hqq_product_faq.cs
@@ -23,7 +23,7 @@ namespace HelloQQPortal.Database
         public Nullable<System.DateTime> modified_on { get; set; }
         public Nullable<int> modified_by { get; set; }
         public sbyte status { get; set; }
-        public sbyte order { get; set; }
+        public int order { get; set; }
     
         public virtual hqq_product hqq_product { get; set; }
     }
